Move http responses to entities package
