+++ postdetails/src/main/java/com/neaniesoft/vermilion/postdetails/data/http/entities/JsonResponse.kt
@@ -1,0 +1,7 @@
+package com.neaniesoft.vermilion.postdetails.data.http.entities
+
+import com.fasterxml.jackson.annotation.JsonProperty
+
+data class JsonResponse(
+    @JsonProperty("data") val data: JsonData
+)
