+++ postdetails/src/main/java/com/neaniesoft/vermilion/postdetails/data/http/entities/MoreCommentsResponse.kt
@@ -1,0 +1,7 @@
+package com.neaniesoft.vermilion.postdetails.data.http.entities
+
+import com.fasterxml.jackson.annotation.JsonProperty
+
+data class MoreCommentsResponse(
+    @JsonProperty("json") val json: JsonResponse
+)
